commit ajustement css

+++ view/forum/listTopics.php
@@ -6,26 +6,33 @@
 <h1 class="bienvenue" >Liste des topics</h1>
 
 <?php
-foreach($topics as $topic ){ ?>
-        <div class="topic"><p class="lien-topic"><a class="lien-list lien-nav" href="index.php?ctrl=forum&action=listPostsByTopic&id=<?php echo $topic->getId() ?>"><h2><?php echo $topic ?></h2></a></p><p class="auteur-topic">publié par <?php echo $topic->getUser() ?> le <?php echo $topic->getDateCreation()?></p></div>
-       
-    <!-- donne droit à l'admin de tout faire sur le site et donne droit à l'auteur de manipuler ses publications -->
-<div class="lien-topic">
-<?php if($topic->getUser()->getId() === App\Session::getUser()->getId() || $_SESSION["user"]->hasRole('ROLE_ADMIN') === APP\Session::isAdmin())
-        { ?>
-            <p class="lien"><a class="lien-gestion1" href="index.php?ctrl=forum&action=deleteTopicById&id=<?= $topic->getId() ?>">Supprimer</a></p>
-            <p class="lien"><a class="lien-gestion2" href="index.php?ctrl=forum&action=updateTopicById&id=<?= $topic->getId() ?>">Modifier</a></p>
-            <!-- si getClosed = 1 alors affiche déverrouiller sinon verrouiller  -->
-<?php if($topic->getClosed())
-      { ?>
-            <p class="lien"><a class="lien-gestion3" href="index.php?ctrl=security&action=unlockTopic&id=<?= $topic->getId() ?>">Déverrouiller<i class="fa-solid fa-lock-open"></i></a></p>
-<?php } else
-      { ?>
-            <p class="lien"><a class="lien-gestion3" href="index.php?ctrl=security&action=lockTopic&id=<?= $topic->getId() ?>">Verrouiller<i class="fa-solid fa-lock"></i></a></p>
-<?php } ?>   
-</div>     
-<?php  }
-}  ?>
+if($topics)
+{
+    foreach($topics as $topic ){ ?>
+            <div class="topic"><p class="lien-topic"><a class="lien-list lien-nav" href="index.php?ctrl=forum&action=listPostsByTopic&id=<?php echo $topic->getId() ?>"><h2 class="bienvenue"><?php echo $topic ?></h2></a></p><p class="auteur-topic">publié par <?php echo $topic->getUser() ?> le <?php echo $topic->getDateCreation()?></p></div>
+           
+        <!-- donne droit à l'admin de tout faire sur le site et donne droit à l'auteur de manipuler ses publications -->
+    <div class="lien-topic">
+    <?php if($topic->getUser()->getId() === App\Session::getUser()->getId() || $_SESSION["user"]->hasRole('ROLE_ADMIN') === APP\Session::isAdmin())
+            { ?> <div class="l1">
+                <p class="lien"><a class="lien-gestion1" href="index.php?ctrl=forum&action=deleteTopicById&id=<?= $topic->getId() ?>">Supprimer</a></p>
+                <p class="lien"><a class="lien-gestion2" href="index.php?ctrl=forum&action=updateTopicById&id=<?= $topic->getId() ?>">Modifier</a></p>
+                <!-- si getClosed = 1 alors affiche déverrouiller sinon verrouiller  -->
+    <?php if($topic->getClosed())
+          { ?>
+                <p class="lien"><a class="lien-gestion3" href="index.php?ctrl=security&action=unlockTopic&id=<?= $topic->getId() ?>">Déverrouiller<i class="fa-solid fa-lock-open"></i></a></p>
+    <?php } else
+          { ?>
+                <p class="lien"><a class="lien-gestion3" href="index.php?ctrl=security&action=lockTopic&id=<?= $topic->getId() ?>">Verrouiller<i class="fa-solid fa-lock"></i></a></p>
+    <?php } ?>  
+    </div> 
+    </div>     
+    <?php  }
+    }  
+}else
+{
+    echo "<h1 class='bienvenue'>Il n'y a aucun topic dans cette catégorie ! </h1>";
+} ?>
 
 <h1 class="bienvenue" >Ajouter un topic</h1>
         <form class="form" action="index.php?ctrl=forum&action=addTopic&id=<?= $category->getId() ?>" method="post">
